util_thread: interrupt thread in stop method
In general we can't be sure, that a non interrupted thread will finish at any time. So calling interrupt before join is absolutly neccessary.

In the future we should remove every state call in a thread against global variables like:

if (global.wantExit)
	this->thread->stop();

this is bad style and should be avoided as much as possible. Interruption point is the way to go, because this way, every single thread is able to shut down without failures. But then every thread is needing well defined interruption points, which will lead to a terminating of the running thread.

This will take effect in a further commit, where the local game sync thread isn't stopped at client quit.

NOTE: this is a canditate for the stable branches