usercore: add interruption point in the LoginThread to abort cleanly game syncing
this will fix #121 partially

without this patch a desurium client quit can take up to several minutes (likely 10).
this path will reduce this time to about 30 seconds max, but it will quit at a well defined state in the databse, which is required to prevent following bugs.

Maybe it is also possible to interrupt the while precedure intermediatly, but on a stable branch I don't want to do too risky things.

NOTE: this is a canditage for the stable branch.